Created taxes package and corrected testing
